requirements.txt
Django==1.9.5
psycopg2==2.6.1
wsgiref==0.1.2
Markdown==2.6.6
djangorestframework==3.3.3
